Downloads Manager: move columns parsing to separate file
